Add files via upload
Streamlined some things, added the ability to check if a voice in the its allowable range, made it so that the chord and the bassline are changed when the soprano voice is impossible to write. Sometimes the soprano voice adds a string instead of an integer. I think this comes from when the bass voice is generated, b/c sometimes it does weird stuff if I print out each bass note as they're written.